Для класса объектов Employee созданы производные классы FixedSalaryEmployee (для фиксированной зарплаты) и HourlySalaryEmployee (для почасовой зарплаты)

+++ EmployeeDataBaseByCompany/EmployeeDataBaseByCompany/Employee.cs
@@ -37,6 +37,6 @@ namespace EmployeeDataBaseByCompany
             return Email_;
         }
 
-        public abstract decimal CalculateSalary();
+        public abstract double CalculateSalary();
     }
 }
